+++ r-application-gateway.tf
@@ -16,6 +16,8 @@ resource "azurerm_application_gateway" "app_gateway" {
 
   zones = var.zones
 
+  firewall_policy_id = var.firewall_policy_id 
+
   enable_http2 = var.enable_http2
 
   frontend_ip_configuration {